Server: cross-gallery stats endpoint (#286, slice 6a)
Sets up the cross-gallery path that GlobalStats will migrate to,
unblocking the deletion of the client-side photo walk. Server
extracts the pure compute logic so both flavours share it.

Refactor:

- `lib/stats-compute.ts` — pure functions lifted out of
  `models/stats.ts`. `computeStats(photos, filter)`, the bucket
  builders, the response shape types, `UNKNOWN_BUCKET`, and
  `isUnfilteredBase` all move here. No DB or cache imports.
- `models/stats.ts` collapses to two thin orchestration calls:
  `getGalleryStats(galleryId, filter?, lang?)` and the new
  `getGlobalStats(filter?, lang?)`. Both wrap the same compute
  with a per-flavour cache key.

New endpoint:

- `POST /api/v1/stats` — admin-only, `requireUnscoped` (rejects
  on hostname-bound instances). Body shape matches the
  gallery-scoped endpoint exactly so the client can reuse the
  same wire types.

Cache:

- Dedicated `:global` / `:global:<lang>` namespace prefix so a
  gallery whose id collides can't poison the cross-gallery
  cache. Slug-shaped gallery ids never start with `:`.
- `invalidateGlobal()` exported alongside `invalidateGallery`.
  `invalidateGallery` upgraded to drop every `<id>` /
  `<id>:<lang>` entry (multi-lang invalidation, drive-by fix —
  previously only the bare `<id>` key was dropped, leaving
  localised variants stale).
- Photo writes (`createPhoto`, `updatePhoto`, `deletePhoto`,
  regeocode via `invalidateStatsForPhoto`) all drop the global
  cache. Gallery-photo link / unlink doesn't — the catalogue set
  is unchanged, only the gallery's membership view shifted.

Client service: `statsService.getGlobalStats(filter, lang)`
matches the gallery method's signature minus the id.

5 new integration tests:
- admin sees the cross-gallery total (5 photos, jp + nl + fi).
- guest blocked (403).
- non-admin blocked (403).
- filter narrows across galleries.
- photo update invalidates the global cache (next call sees the
  new country bucket).

No frontend changes yet — slice 6b migrates GlobalStats off
`stats.generate` to consume the new endpoint and deletes the
client-side photo walk, closing #286.
